Fixed Login payload, takes DB credentials now (CHECK DESCRIPTION FOR NEW LOGIN CREDENTIALS)
CURRENT USERNAME: admin
CURRENT PASSWORD: testing

took out the API call in login.vue, since loggedInUser.ts was already doing the same thing

previously, the payload from frontend to backend was being sent as "admin" and "admin" for username and password, which was fine because that's what they both were. However, I tried to test changing the password, and it was still sending "admin" and "admin", which meant there was a fundamental bug with the code.

This current implementation reads the password in from the DB, so it's not hard coded.

TO-DO: implement bcrypt encryption and password checker
(Will replace "testing" in DB with an actual hash)