Created a temporary password changer for client to use
I'm putting it in the backend folder because it uses the Credentials.py and SQL.py files.
I am also assuming that it's somewhat "safe" to put this password changer in this folder, because the Credentials.py file is also in here.

If the creds file isn't secure, I think this password changer would be the least of our problems.

This file is a temporary fix in case the password is forgotten